requirements.txt
faster_whisper==1.0.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cudnn-cu11==8.9.6.50
gputil
pynvml
pydub
google-cloud-storage==2.16.0